Improve tox config
To make it easier to add testing dependencies without having to modify
the tox.ini file we make tox install from the requirements-dev.txt file.
We also move flake to a seperate testing environment so it can be run
independent of the tests. We also run pytest with the strict option so
markers are required to be declared before use.